bubble_board_dashboard/requirements.txt
streamlit==1.38.0
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
requests==2.32.3
yfinance==0.2.43
